models working properly, fixed maven dependencies for prersistence, added classes to persistence xml

+++ persistence/src/main/java/repositories/BorrowerRepository.java
@@ -5,7 +5,7 @@ import models.Borrower;
 import javax.persistence.EntityManager;
 import java.math.BigInteger;
 
-public class BorrowerRepository extends GenericRepository<Borrower, BigInteger> {
+public class BorrowerRepository extends GenericRepository<Borrower, Integer> {
     protected BorrowerRepository(EntityManager entityManager) {
         super(entityManager);
     }
